requirements.txt
vkbottle
google-genai
python-dotenv
httpx[socks]